feat/#15: 수동입력뷰 사이즈 이슈 해결, 리액트 버전 18 > 17로 다운그레이드, 파일 구조 원상복구
