+++ package.json
@@ -14,8 +14,8 @@
   "dependencies": {
     "@types/styled-components": "^5.1.26",
     "axios": "^1.2.1",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "recoil": "^0.7.6",
     "styled-components": "^5.3.6"
   },
