+++ src/components/common/Header/Header.tsx
@@ -22,6 +22,13 @@ function Header(props: HeaderProps) {
   };
 
   const goBackToHistory = () => {
+    document.body.style.width = '38rem';
+    document.body.style.height = '37.5rem';
+    const container = document.getElementById('app-container') as HTMLElement;
+    if (container) {
+      container.style.width = '38rem';
+      container.style.height = '37.5rem';
+    }
     history && setCurrentView(history);
   };
