+++ src/manifest.json
@@ -17,7 +17,7 @@
   "content_scripts": [
     {
       "matches": ["https://www.musinsa.com/*"],
-      "js": ["script/sizeTableContent.js", "script/productContent.js", "script/injectContent.js"]
+      "js": ["script/sizeTableContent.js", "script/productContent.js"]
     }
   ],
   "externally_connectable": {
